week2_app/requirements.txt
gradio==4.19.2
Pillow==9.4.0
Pillow==10.2.0
torch==2.2.1+cu118
transformers==4.38.1
